+++ .storybook/webpack.config.js
@@ -1,0 +1,3 @@
+// For more information refer the docs: https://storybook.js.org/configurations/custom-webpack-config
+
+module.exports = ({ config }) => config