+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   very basic docs site: https://at0g.github.io/hooks/
+-   storybook example
+-   useMedia
 
 ## [0.0.3] - 2019-11-02